CPU: set_addr_mode() is semi complete!
Currently it is untested but built to where I am happy with it
added a temporary u8 ROM[] array to act as the program rom to
allow for the program to compile. I could make it just slurp up some
file pointer but I might want it to be a fully flushed class instead
but this might work.